moved config singleton to config.py file
